requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.2.0=pypi_0
ca-certificates=2023.11.17=h56e8100_0
certifi=2023.11.17=pypi_0
chardet=3.0.4=pypi_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pypi_0
dalex=1.6.0=pypi_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.7=pypi_0
entrypoints=0.4=pypi_0
gluonts=0.6.7=pypi_0
holidays=0.9.12=pypi_0
idna=2.6=pypi_0
intel-openmp=2023.2.0=h57928b3_50497
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
jedi=0.19.1=pypi_0
joblib=1.3.2=pyhd8ed1ab_0
jupyter-client=7.4.9=pypi_0
jupyter-core=4.12.0=pypi_0
kiwisolver=1.4.5=pypi_0
libblas=3.9.0=20_win64_mkl
libcblas=3.9.0=20_win64_mkl
liblapack=3.9.0=20_win64_mkl
libsqlite=3.44.2=hcfcfb64_0
lightgbm=4.1.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.2.0=h6a75c08_50497
msys2-conda-epoch=20160418=1
mxnet=1.7.0.post1=pypi_0
nest-asyncio=1.5.8=pypi_0
numpy=1.21.6=pypi_0
openssl=3.2.0=hcfcfb64_1
packaging=23.2=pypi_0
pandas=1.2.5=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
plotly=5.18.0=pypi_0
prompt-toolkit=3.0.41=pypi_0
psutil=5.9.6=pypi_0
pydantic=1.6.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.7.12=h900ac77_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-graphviz=0.8.4=pypi_0
python_abi=3.7=4_cp37m
pytz=2023.3.post1=pyhd8ed1ab_0
pywin32=306=pypi_0
pyzmq=25.1.2=pypi_0
requests=2.18.4=pypi_0
scikit-learn=0.22.2.post1=py37h7208079_0
scipy=1.7.3=pypi_0
setuptools=59.8.0=py37h03978a9_1
six=1.16.0=pyh6c4a22f_0
sqlite=3.44.2=hcfcfb64_0
statsmodels=0.13.5=pypi_0
tbb=2021.7.0=h91493d7_0
tenacity=8.2.3=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.7.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
ujson=1.35=pypi_0
urllib3=1.22=pypi_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
wcwidth=0.2.12=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
polars = 0.18.4s
pykalman == 0.9.7